+++ src/subject.test.ts
@@ -6,12 +6,12 @@ import { createDocument } from './document';
 import { rdf } from 'rdf-namespaces';
 
 const mockDocument = 'https://document.com/';
-const mockSubjectWithLiteralThenNode = 'https://subject1.com/';
-const mockSubjectWithNodeThenLiteral = 'https://subject2.com/';
+const mockSubjectWithLiteralThenRef = 'https://subject1.com/';
+const mockSubjectWithRefThenLiteral = 'https://subject2.com/';
 const mockSubjectWithLiteral = 'https://subject3.com/';
-const mockSubjectWithNode = 'https://subject4.com/';
+const mockSubjectWithRef = 'https://subject4.com/';
 const mockSubjectWithTwoLiterals = 'https://subject5.com/';
-const mockSubjectWithTwoNodes = 'https://subject6.com/';
+const mockSubjectWithTwoRefs = 'https://subject6.com/';
 const mockSubjectWithDateLiteral = 'https://subject7.com/';
 const mockSubjectWithIntegerLiteral = 'https://subject8.com/';
 const mockSubjectWithDecimalLiteral = 'https://subject9.com/';
@@ -19,8 +19,8 @@ const mockSubjectWithDifferentTypesOfLiterals = 'https://subject10.com/';
 const mockTypedSubject = 'https://subject7.com/';
 const mockEmptySubject = 'https://empty-subject.com/';
 const mockPredicate = 'https://mock-predicate.com/';
-const mockObjectNode = 'https://mock-object.com/';
-const mockObjectNode2 = 'https://mock-object-2.com/';
+const mockObjectRef = 'https://mock-object.com/';
+const mockObjectRef2 = 'https://mock-object-2.com/';
 const mockLiteralValue = 'Arbitrary literal value';
 const mockObjectLiteral = lit(mockLiteralValue, null as any, null as any);
 const mockLiteralValue2 = 'Another arbitrary literal value';
@@ -33,18 +33,18 @@ const mockLiteralDecimal = 4.2;
 const mockObjectDecimalLiteral = Literal.fromNumber(mockLiteralDecimal);
 const mockTypeObject = 'https://mock-type-object.com/';
 const mockStatements = [
-  st(sym(mockSubjectWithLiteralThenNode), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
-  st(sym(mockSubjectWithLiteralThenNode), sym(mockPredicate), sym(mockObjectNode), sym(mockDocument)),
-  st(sym(mockSubjectWithNodeThenLiteral), sym(mockPredicate), sym(mockObjectNode), sym(mockDocument)),
-  st(sym(mockSubjectWithNodeThenLiteral), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
+  st(sym(mockSubjectWithLiteralThenRef), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
+  st(sym(mockSubjectWithLiteralThenRef), sym(mockPredicate), sym(mockObjectRef), sym(mockDocument)),
+  st(sym(mockSubjectWithRefThenLiteral), sym(mockPredicate), sym(mockObjectRef), sym(mockDocument)),
+  st(sym(mockSubjectWithRefThenLiteral), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
   st(sym(mockSubjectWithLiteral), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
-  st(sym(mockSubjectWithNode), sym(mockPredicate), sym(mockObjectNode), sym(mockDocument)),
+  st(sym(mockSubjectWithRef), sym(mockPredicate), sym(mockObjectRef), sym(mockDocument)),
   st(sym(mockSubjectWithTwoLiterals), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
   st(sym(mockSubjectWithTwoLiterals), sym(mockPredicate), mockObjectLiteral2, sym(mockDocument)),
-  st(sym(mockSubjectWithTwoLiterals), sym(mockPredicate), sym(mockObjectNode), sym(mockDocument)),
-  st(sym(mockSubjectWithTwoNodes), sym(mockPredicate), sym(mockObjectNode), sym(mockDocument)),
-  st(sym(mockSubjectWithTwoNodes), sym(mockPredicate), sym(mockObjectNode2), sym(mockDocument)),
-  st(sym(mockSubjectWithTwoNodes), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
+  st(sym(mockSubjectWithTwoLiterals), sym(mockPredicate), sym(mockObjectRef), sym(mockDocument)),
+  st(sym(mockSubjectWithTwoRefs), sym(mockPredicate), sym(mockObjectRef), sym(mockDocument)),
+  st(sym(mockSubjectWithTwoRefs), sym(mockPredicate), sym(mockObjectRef2), sym(mockDocument)),
+  st(sym(mockSubjectWithTwoRefs), sym(mockPredicate), mockObjectLiteral, sym(mockDocument)),
   st(sym(mockTypedSubject), sym(rdf.type), sym(mockTypeObject), sym(mockDocument)),
   st(sym(mockSubjectWithDateLiteral), sym(mockPredicate), mockObjectDateTimeLiteral, sym(mockDocument)),
   st(sym(mockSubjectWithDecimalLiteral), sym(mockPredicate), mockObjectDecimalLiteral, sym(mockDocument)),
@@ -66,11 +66,11 @@ function getMockTripleDocument() {
   return mockTripleDocument;
 }
 
-describe('asNodeRef', () => {
+describe('asRef', () => {
   it('should give access to the IRI that represents this Subject', () => {
     const mockTripleDocument = getMockTripleDocument();
     const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
-    expect(subject.asNodeRef())
+    expect(subject.asRef())
       .toBe(mockSubjectWithLiteral);
   });
 });
@@ -123,9 +123,9 @@ describe('getLiteral', () => {
       .toEqual(mockLiteralDateTime.getTime());
   });
 
-  it('should return null if a Node is found instead of a Literal', () => {
+  it('should return null if a Reference is found instead of a Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
     expect(subject.getLiteral(mockPredicate))
       .toBeNull();
   });
@@ -139,14 +139,14 @@ describe('getLiteral', () => {
 
   it('should return the first found value if that is a Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
     expect(subject.getLiteral(mockPredicate))
       .toBe(mockLiteralValue);
   });
 
   it('should return the second found value if that is the first Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNodeThenLiteral);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRefThenLiteral);
     expect(subject.getLiteral(mockPredicate))
       .toBe(mockLiteralValue);
   });
@@ -168,9 +168,9 @@ describe('getString', () => {
   });
 
 
-  it('should return null if a Node is found instead of a string Literal', () => {
+  it('should return null if a Reference is found instead of a string Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
     expect(subject.getString(mockPredicate))
       .toBeNull();
   });
@@ -192,9 +192,9 @@ describe('getInteger', () => {
   });
 
 
-  it('should return null if a Node is found instead of a integer Literal', () => {
+  it('should return null if a Reference is found instead of a integer Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
     expect(subject.getInteger(mockPredicate))
       .toBeNull();
   });
@@ -216,9 +216,9 @@ describe('getDecimal', () => {
   });
 
 
-  it('should return null if a Node is found instead of a decimal Literal', () => {
+  it('should return null if a Reference is found instead of a decimal Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
     expect(subject.getDecimal(mockPredicate))
       .toBeNull();
   });
@@ -240,9 +240,9 @@ describe('getDate', () => {
   });
 
 
-  it('should return null if a Node is found instead of a Date Literal', () => {
+  it('should return null if a Reference is found instead of a Date Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
     expect(subject.getDateTime(mockPredicate))
       .toBeNull();
   });
@@ -335,47 +335,47 @@ describe('getAllLiterals', () => {
   });
 });
 
-describe('getNodeRef', () => {
-  it('should return a found Node reference', () => {
+describe('getRef', () => {
+  it('should return a found Reference', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
-    expect(subject.getNodeRef(mockPredicate))
-      .toEqual(mockObjectNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
+    expect(subject.getRef(mockPredicate))
+      .toEqual(mockObjectRef);
   });
 
-  it('should return null if a Node is found instead of a Literal', () => {
+  it('should return null if a Reference is found instead of a Literal', () => {
     const mockTripleDocument = getMockTripleDocument();
     const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteral);
-    expect(subject.getNodeRef(mockPredicate))
+    expect(subject.getRef(mockPredicate))
       .toBeNull();
   });
 
   it('should return null if nothing is found', () => {
     const mockTripleDocument = getMockTripleDocument();
     const subject = initialiseSubject(mockTripleDocument, mockEmptySubject);
-    expect(subject.getNodeRef(mockPredicate))
+    expect(subject.getRef(mockPredicate))
       .toBeNull();
   });
 
-  it('should return the first found value if that is a Node', () => {
+  it('should return the first found value if that is a Reference', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNodeThenLiteral);
-    expect(subject.getNodeRef(mockPredicate))
-      .toBe(mockObjectNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRefThenLiteral);
+    expect(subject.getRef(mockPredicate))
+      .toBe(mockObjectRef);
   });
 
-  it('should return the second found value if that is the first Node', () => {
+  it('should return the second found value if that is the first Reference', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenNode);
-    expect(subject.getNodeRef(mockPredicate))
-      .toBe(mockObjectNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
+    expect(subject.getRef(mockPredicate))
+      .toBe(mockObjectRef);
   });
 
-  it('should only return a single Node', () => {
+  it('should only return a single Reference', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithTwoNodes);
-    expect(subject.getNodeRef(mockPredicate))
-      .toBe(mockObjectNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithTwoRefs);
+    expect(subject.getRef(mockPredicate))
+      .toBe(mockObjectRef);
   });
 });
 
@@ -393,18 +393,18 @@ describe('getType', () => {
   });
 });
 
-describe('getAllNodeRefs', () => {
-  it('should only return Nodes', () => {
+describe('getAllRefs', () => {
+  it('should only return References', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithTwoNodes);
-    expect(subject.getAllNodeRefs(mockPredicate))
-      .toEqual([mockObjectNode, mockObjectNode2]);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithTwoRefs);
+    expect(subject.getAllRefs(mockPredicate))
+      .toEqual([mockObjectRef, mockObjectRef2]);
   });
 
   it('should return an empty array if nothing is found', () => {
     const mockTripleDocument = getMockTripleDocument();
     const subject = initialiseSubject(mockTripleDocument, mockEmptySubject);
-    expect(subject.getAllNodeRefs(mockPredicate))
+    expect(subject.getAllRefs(mockPredicate))
       .toEqual([]);
   });
 });
@@ -517,23 +517,23 @@ describe('removeLiteral', () => {
 });
 
 describe('setLiteral', () => {
-  it('should remove all existing values, whether Literal or NodeRef', () => {
+  it('should remove all existing values, whether Literal or Reference', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
     subject.setLiteral(mockPredicate, mockLiteralValue2);
     const [pendingDeletions, pendingAdditions] = subject.getPendingStatements();
     expect(pendingDeletions).toEqual([
       st(
-        sym(mockSubjectWithLiteralThenNode),
+        sym(mockSubjectWithLiteralThenRef),
         sym(mockPredicate),
         mockObjectLiteral,
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockTripleDocument.asRef()),
       ),
       st(
-        sym(mockSubjectWithLiteralThenNode),
+        sym(mockSubjectWithLiteralThenRef),
         sym(mockPredicate),
-        sym(mockObjectNode),
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockObjectRef),
+        sym(mockTripleDocument.asRef()),
       ),
     ]);
     expect(pendingAdditions.length).toBe(1);
@@ -542,89 +542,89 @@ describe('setLiteral', () => {
   });
 });
 
-describe('addNodeRef', () => {
+describe('addRef', () => {
   it('should produce Statements that the Document can store in the user\'s Pod', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
-    subject.addNodeRef(mockPredicate, mockObjectNode2);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
+    subject.addRef(mockPredicate, mockObjectRef2);
     const [pendingDeletions, pendingAdditions] = subject.getPendingStatements();
     expect(pendingDeletions).toEqual([]);
     expect(pendingAdditions)
       .toEqual([st(
-        sym(mockSubjectWithNode),
+        sym(mockSubjectWithRef),
         sym(mockPredicate),
-        sym(mockObjectNode2),
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockObjectRef2),
+        sym(mockTripleDocument.asRef()),
       )]);
   });
 });
 
-describe('removeNodeRef', () => {
+describe('removeRef', () => {
   it('should produce Statements that the Document can apply to the user\'s Pod', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithNode);
-    subject.removeNodeRef(mockPredicate, mockObjectNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithRef);
+    subject.removeRef(mockPredicate, mockObjectRef);
     const [pendingDeletions, pendingAdditions] = subject.getPendingStatements();
     expect(pendingAdditions).toEqual([]);
     expect(pendingDeletions)
       .toEqual([st(
-        sym(mockSubjectWithNode),
+        sym(mockSubjectWithRef),
         sym(mockPredicate),
-        sym(mockObjectNode),
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockObjectRef),
+        sym(mockTripleDocument.asRef()),
       )]);
   });
 });
 
-describe('setNodeRef', () => {
-  it('should remove all existing values, whether Literal or NodeRef', () => {
+describe('setRef', () => {
+  it('should remove all existing values, whether Literal or Ref', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenNode);
-    subject.setNodeRef(mockPredicate, mockObjectNode2);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
+    subject.setRef(mockPredicate, mockObjectRef2);
     const [pendingDeletions, pendingAdditions] = subject.getPendingStatements();
     expect(pendingDeletions).toEqual([
       st(
-        sym(mockSubjectWithLiteralThenNode),
+        sym(mockSubjectWithLiteralThenRef),
         sym(mockPredicate),
         mockObjectLiteral,
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockTripleDocument.asRef()),
       ),
       st(
-        sym(mockSubjectWithLiteralThenNode),
+        sym(mockSubjectWithLiteralThenRef),
         sym(mockPredicate),
-        sym(mockObjectNode),
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockObjectRef),
+        sym(mockTripleDocument.asRef()),
       ),
     ]);
     expect(pendingAdditions)
       .toEqual([st(
-        sym(mockSubjectWithLiteralThenNode),
+        sym(mockSubjectWithLiteralThenRef),
         sym(mockPredicate),
-        sym(mockObjectNode2),
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockObjectRef2),
+        sym(mockTripleDocument.asRef()),
       )]);
   });
 });
 
 describe('removeAll', () => {
-  it('should remove all existing values, whether Literal or NodeRef', () => {
+  it('should remove all existing values, whether Literal or Reference', () => {
     const mockTripleDocument = getMockTripleDocument();
-    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenNode);
+    const subject = initialiseSubject(mockTripleDocument, mockSubjectWithLiteralThenRef);
     subject.removeAll(mockPredicate);
     const [pendingDeletions, pendingAdditions] = subject.getPendingStatements();
     expect(pendingAdditions).toEqual([]);
     expect(pendingDeletions).toEqual([
       st(
-        sym(mockSubjectWithLiteralThenNode),
+        sym(mockSubjectWithLiteralThenRef),
         sym(mockPredicate),
         mockObjectLiteral,
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockTripleDocument.asRef()),
       ),
       st(
-        sym(mockSubjectWithLiteralThenNode),
+        sym(mockSubjectWithLiteralThenRef),
         sym(mockPredicate),
-        sym(mockObjectNode),
-        sym(mockTripleDocument.asNodeRef()),
+        sym(mockObjectRef),
+        sym(mockTripleDocument.asRef()),
       ),
     ]);
   });